Resolved review points.
